requirements.txt
exaroton==0.0.7
dotenv==0.9.9
discord.py==2.5.2
flask==3.1.1